JournalRecord uint32

+++ Main/JournalRecord.cpp
@@ -3,31 +3,31 @@
 using namespace std;
 
 //============================================================
-JournalRecord::JournalRecord(uint64_t tid, bool record_type){
+JournalRecord::JournalRecord(uint32_t tid, bool record_type){
 
     transaction_id = tid;
 	type = record_type;
-	values = new DArray<uint64_t>();
+	values = new DArray<uint32_t>();
     // cout << "New Journal Record created" << endl;
 }
 //============================================================
-void JournalRecord::addValue(uint64_t val){
+void JournalRecord::addValue(uint32_t val){
     // cout << "Adding value: " << val << endl;
     values->push_back(val);
 }
 //============================================================
-uint64_t JournalRecord::getValue(uint64_t key){
+uint32_t JournalRecord::getValue(uint32_t key){
     return values->get(key);
 }
 bool JournalRecord::getType(){
 	return type;
 }
 //============================================================
-DArray<uint64_t>* JournalRecord::getAllValues(){
+DArray<uint32_t>* JournalRecord::getAllValues(){
 	return values;
 }
 //============================================================
-uint64_t JournalRecord::getTransactionId(){
+uint32_t JournalRecord::getTransactionId(){
     return transaction_id;
 }
 //============================================================
